+++ src/main/java/com/b1n_ry/yigd/util/YigdCommands.java
@@ -8,6 +8,7 @@ import com.b1n_ry.yigd.data.ListMode;
 import com.b1n_ry.yigd.networking.LightGraveData;
 import com.b1n_ry.yigd.networking.LightPlayerData;
 import com.b1n_ry.yigd.networking.ServerPacketHandler;
+import com.mojang.authlib.GameProfile;
 import com.mojang.brigadier.context.CommandContext;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
@@ -122,7 +123,7 @@ public class YigdCommands {
     private static int viewAll(CommandContext<ServerCommandSource> context) {
         ServerPlayerEntity player = context.getSource().getPlayer();
 
-        Map<ProfileComponent, List<GraveComponent>> players = DeathInfoManager.INSTANCE.getPlayerGraves();
+        Map<GameProfile, List<GraveComponent>> players = DeathInfoManager.INSTANCE.getPlayerGraves();
 
         List<LightPlayerData> lightPlayerData = new ArrayList<>();
         players.forEach((profile, components) -> {
@@ -134,7 +135,7 @@ public class YigdCommands {
                     case DESTROYED -> destroyed++;
                 }
             }
-            LightPlayerData lightData = new LightPlayerData(components.size(), unclaimed, destroyed, profile);
+            LightPlayerData lightData = new LightPlayerData(components.size(), unclaimed, destroyed, new ProfileComponent(profile));
             lightPlayerData.add(lightData);
         });
 
@@ -244,11 +245,11 @@ public class YigdCommands {
     }
     private static int showList(CommandContext<ServerCommandSource> context) {
         ListMode listMode = DeathInfoManager.INSTANCE.getGraveListMode();
-        Set<ProfileComponent> affectedPlayers = DeathInfoManager.INSTANCE.getAffectedPlayers();
+        Set<GameProfile> affectedPlayers = DeathInfoManager.INSTANCE.getAffectedPlayers();
 
         StringJoiner joiner = new StringJoiner(", ");
-        for (ProfileComponent profile : affectedPlayers) {
-            joiner.add(profile.name().orElse("PLAYER_NOT_FOUND"));
+        for (GameProfile profile : affectedPlayers) {
+            joiner.add(Optional.ofNullable(profile.getName()).orElse("PLAYER_NOT_FOUND"));
         }
         context.getSource().sendMessage(Text.literal(listMode.name() + ": %s" + joiner));